+++ lib/plugins/aws/invoke.js
@@ -3,9 +3,9 @@
 const BbPromise = require('bluebird');
 const chalk = require('chalk');
 const path = require('path');
-const validate = require('../lib/validate');
+const validate = require('./lib/validate');
 const stdin = require('get-stdin');
-const formatLambdaLogEvent = require('../utils/formatLambdaLogEvent');
+const formatLambdaLogEvent = require('./utils/formatLambdaLogEvent');
 
 class AwsInvoke {
   constructor(serverless, options) {
