+++ lib/plugins/index.js
@@ -25,7 +25,7 @@ module.exports = [
   require('./aws/common/index.js'),
   require('./aws/package/index.js'),
   require('./aws/deploy/index.js'),
-  require('./aws/invoke/index.js'),
+  require('./aws/invoke.js'),
   require('./aws/info/index.js'),
   require('./aws/logs/index.js'),
   require('./aws/metrics/awsMetrics.js'),
